+++ CHANGELOG.md
@@ -25,6 +25,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 - New `WildcardType` enum [#853]
 - New `DescriptorPublicKey::add_wildcard` method, which adds an unhardened wildcard to the derivation path of the descriptor [#853]
 - New `DescriptorSecretKey::add_wildcard(wildcard_type: WildcardType)` method, which adds a wildcard to the derivation path of the descriptor [#853]
+- Exposed `new_sh`, `new_wsh`,`new_bare` and `new_sh_wsh` methods on `Descriptor` type [#988]
 
 [#853]: https://github.com/bitcoindevkit/bdk-ffi/pull/853
 [#853]: https://github.com/bitcoindevkit/bdk-ffi/pull/945
@@ -32,6 +33,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 [#986]: https://github.com/bitcoindevkit/bdk-ffi/pull/971
 [#986]: https://github.com/bitcoindevkit/bdk-ffi/pull/973
 [#986]: https://github.com/bitcoindevkit/bdk-ffi/pull/986
+[#986]: https://github.com/bitcoindevkit/bdk-ffi/pull/988
 
 ## [v2.3.0]